Change validation call and add custom callback
Change the way the validation functions are called. As the process was started to many times with jQuery 3.3.1. 

Added a callback function to handle the errors and moved the default error handling to a function. So you can handle the errors with a custom function from an other Extension if you want.